pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "leakcheck"
version = "0.1.0"
description = "Leak Check - privacy risk auditing for LLMs"
requires-python = ">=3.10,<3.14"
dependencies = [
  "pydantic>=2.6",
  "pyyaml>=6.0",
  "rich>=13.0",
  "requests>=2.31",
  "python-dotenv>=1.0",
  "numpy>=1.26",
  "torch==2.7.0",
  "torchvision==0.22.0",
  "torchaudio==2.7.0",
  "sentence-transformers>=2.6",
  "pyarrow>=14.0",
  "flask>=3.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.4"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[project.scripts]
leakcheck = "leakcheck.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"
